Big Commit
Forgot to commit before starting on a new task. So i've set up the profile page and the carousel and a highlighted item on the index page. Started to set up the initial amount of credits you have on registration. As of now, i've only setup the credits display in the header and on the profile.